requirements.txt
fastapi
mangum
dotenv
pydantic_settings
uvicorn[standard]
requests
python-jose[cryptography]
boto3
botocore[crt]
httpx